requirements.txt
# A requirements file in python is a file that lists all the libraries that your project depends on.
# it is common to include in a project, so someone can easily install the needed
# libraries to run your project.  
# For example
# pip install -r requirements.txt
# will install all the libraries listed in the requirements.txt file

# Here are some example libaries that someone may include, one is included on each line
Datetime
Django~=4.2.7
requests~=2.31.0
plotly~=5.18.0
numpy~=1.26.2
python-dateutil~=2.8.2


